Add code to make war exceutable on 3.0 servers

+++ src/main/java/com/app/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.app.repository;
 
-import com.app.persistence.Customer;
+import com.app.domain.Customer;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +10,7 @@ import java.util.List;
  * Created by ruben on 4/03/16.
  */
 @Repository
-public interface CustomerRepository extends CrudRepository<Customer, Long> {
+public interface CustomerRepository extends CrudRepository<Customer, Integer> {
 
     List<Customer> findByLastName(String lastName);
 }
