+++ src/main/java/com/app/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.app.repository;
 
-import com.app.persistence.Orders;
+import com.app.domain.Orders;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +10,7 @@ import java.util.List;
  * Created by ruben on 4/03/16.
  */
 @Repository
-public interface OrderRepository extends CrudRepository<Orders, Long> {
+public interface OrderRepository extends CrudRepository<Orders, Integer> {
 
     List<Orders> findByName(String lastName);
 }
